# docs/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "High Street Gym ",
    "description": "JSON REST API for Dynamic Gym website "
  },
  "servers": [
    {
      "url": "http://localhost:8090"
    }
  ],
  "basePath": "",
  "paths": {
...
    "/room/upload/xml": {
      "post": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/room": {
      "post": {
        "summary": "Create a room",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Adding new room",
          "content": {
            "application/json": {
              "schema": {
                "room_location": "string",
                "room_number": "number"
              },
              "example": {
                "room_location": "Block A",
                "room_number": 205
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get all room",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "summary": "Update an existing room",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Update an existing room",
          "content": {
            "application/json": {
              "schema": {
                "room_location": "string",
                "room_number": "number",
                "room_id": "number"
              },
              "example": {
                "room_location": "Block G",
                "room_number": 303,
                "room_id": 1
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a specific room by ID",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Deleting a room",
          "content": {
            "application/json": {
              "schema": {
                "room_id": "number"
              },
              "example": {
                "room_id": "1"
              }
            }
          }
        }
      }
    },
    "/room/{id}": {
      "get": {
        "summary": "Get a specific room by ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
...
}